Task

+++ src/AnimalRescue/Animal.java
@@ -4,7 +4,7 @@ public class Animal {
 
 
     private String breed ;
-    private byte age ;
+    private int age ;
     private float weight ;
     private int heightInCm ;
     private String name ;
